feat(transactions): Add logic to uptade transactions with pending status

+++ back/src/api/models/Transaction.js
@@ -104,11 +104,11 @@ Transaccion.init({
         allowNull: false
     },
     amount: {
-        type: DataTypes.DECIMAL(10, 2),
+        type: DataTypes.INTEGER,
         allowNull: false
     },
     commission: {
-        type: DataTypes.DECIMAL(10, 2),
+        type: DataTypes.INTEGER,
         allowNull: false
     },
     createdAt: {
